create wrk base class
